pyproject.toml
[project]
name = "woodburn"
version = "0.1.0"
description = "Woodburn Website"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "gunicorn>=23.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "authlib>=1.3.0",
    "flask-caching>=2.3.1",
]

[dependency-groups]
dev = [
    "pre-commit>=4.4.0",
    "ruff>=0.14.5",
]
